workshops/BH Asia 2024 Arsenal Lab - Catsniffer/Files/pycatsniffer_bv3/requirements.txt
click==8.1.7
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.17.2
pyserial==3.5
rich==13.7.0
typer==0.9.0
typing_extensions==4.9.0
